Add support for removed date key in depcrecated section
*  Add removed_at_date key in deprecated section
*  removed_at_date value is 2 years from current date
*  Remove removed_version key from deprecated section
*  FIx short_description update issue
*  Move common code to utils file
